Update buildscripts to most recent - includes some hardcodes for my old builders. I'll deal with those next.

+++ build-scripts/import-databases.sh
@@ -6,6 +6,7 @@
 set -e
 #set -x
 
+[[ -n "$NO_DB_IMPORT" ]] && exit 0
 
 SqlCvsCred=""
 if [ "$sqlcvs_host" ] ; then SqlCvsCred="$SqlCvsCred -h$sqlcvs_host"; fi
@@ -19,6 +20,10 @@ if [ "$sql_build_port" ] ; then BuildCred="$BuildCred -P$sql_build_port"; fi
 if [ "$sql_build_user" ] ; then BuildCred="$BuildCred -u$sql_build_user"; fi
 if [ "$sql_build_pass" ] ; then BuildCred="$BuildCred -p$sql_build_pass"; fi
 
+if [[ "$DB_IMPORT" == "no" ]] ; then
+	DisplayMessage "Database import disabled - Skipping"
+	exit 0;
+fi
 
 function get_dbdump 
 {
@@ -29,7 +34,7 @@ function get_dbdump
 	DisplayMessage "Getting a sqldump of ${dbdump_name}"
 
 #	[ ! -e "${dbdump_file}.prep" ] && #ZaercTempHack
-	mysqldump --compress --skip-extended-insert $SqlCvsCred "${dbdump_name}" > "${dbdump_file}.prep"
+	mysqldump --column-statistics=0 --compress --skip-extended-insert $SqlCvsCred "${dbdump_name}" > "${dbdump_file}.prep"
 
 	cp -f "${dbdump_file}.prep" "${dbdump_file}"
 
@@ -61,49 +66,51 @@ function import_databases () {
 
 	mkdir -p "${dbdump_dir}"
 
-
 	## Import pluto_main, pluto_media, pluto_security, pluto_telecom
 	get_dbdump main_sqlcvs_utf8 "$dbdump_main_sqlcvs"
-	get_dbdump myth_sqlcvs "$dbdump_myth_sqlcvs"
-	get_dbdump pluto_media "$dbdump_pluto_media"
-	get_dbdump pluto_security "$dbdump_pluto_security"
-	get_dbdump pluto_telecom "$dbdump_pluto_telecom"
-	get_dbdump lmce_game "$dbdump_lmce_game"
+	if [ -z "$IMPORT_BUILD_DB_ONLY" ]; then
+		get_dbdump myth_sqlcvs "$dbdump_myth_sqlcvs"
+		get_dbdump pluto_media "$dbdump_pluto_media"
+		get_dbdump pluto_security "$dbdump_pluto_security"
+		get_dbdump pluto_telecom "$dbdump_pluto_telecom"
+		get_dbdump lmce_game "$dbdump_lmce_game"
+	fi
 
 	# Run search and replace over the db dumps
 	DisplayMessage "Running Search'n'Replace on the sqldumps (MakeRelase_PrepFiles)"
-	export LD_LIBRARY_PATH="$mkr_dir:${svn_dir}/${svn_branch_name}/src/lib"
+	export LD_LIBRARY_PATH="$mkr_dir:${scm_dir}/src/lib"
 	MakeRelease_PrepFiles="${mkr_dir}/MakeRelease_PrepFiles"
 	$MakeRelease_PrepFiles -p $(dirname $dbdump_main_sqlcvs) -e "$(basename $dbdump_main_sqlcvs),$(basename $dbdump_myth_sqlcvs),$(basename $dbdump_pluto_media),$(basename $dbdump_pluto_security),$(basename $dbdump_pluto_telecom)" -c /etc/lmce-build/${flavor}.conf
 
-	
+
 	# pluto_main_build & pluto_main
 	import_dbdump pluto_main_build "$dbdump_main_sqlcvs"
-	import_dbdump pluto_main "$dbdump_main_sqlcvs"
+	if [ -z "$IMPORT_BUILD_DB_ONLY" ]; then
+		import_dbdump pluto_main "$dbdump_main_sqlcvs"
 
-	# main_sqlcvs_ubuntu
-	import_dbdump "main_sqlcvs_${flavor}" "$dbdump_main_sqlcvs"
-	mysql $BuildCred "main_sqlcvs_${flavor}" -e 'DELETE FROM `Package_Version`;'
-	mysql $BuildCred "main_sqlcvs_${flavor}" -e 'DELETE FROM `Schema`;'
+		# main_sqlcvs_ubuntu
+		import_dbdump "main_sqlcvs_${flavor}" "$dbdump_main_sqlcvs"
+		mysql $BuildCred "main_sqlcvs_${flavor}" -e 'DELETE FROM `Package_Version`;'
+		mysql $BuildCred "main_sqlcvs_${flavor}" -e 'DELETE FROM `Schema`;'
 
-	# myth_sqlcvs_ubuntu
-	import_dbdump "myth_sqlcvs_${flavor}" "$dbdump_myth_sqlcvs"
+		# myth_sqlcvs_ubuntu
+		import_dbdump "myth_sqlcvs_${flavor}" "$dbdump_myth_sqlcvs"
 
-	# pluto_media
-	import_dbdump pluto_media "$dbdump_pluto_media"
+		# pluto_media
+		import_dbdump pluto_media "$dbdump_pluto_media"
 
-	# pluto_security
-	import_dbdump pluto_security "$dbdump_pluto_security"
+		# pluto_security
+		import_dbdump pluto_security "$dbdump_pluto_security"
 
-	# pluto_telecom
-	import_dbdump pluto_telecom "$dbdump_pluto_telecom"
+		# pluto_telecom
+		import_dbdump pluto_telecom "$dbdump_pluto_telecom"
 
-	# lmce_game
-	import_dbdump lmce_game "$dbdump_lmce_game"
+		# lmce_game
+		import_dbdump lmce_game "$dbdump_lmce_game"
 
-#	rm -f "$dbdump_main_sqlcvs" "$dbdump_myth_sqlcvs" "$dbdump_pluto_media" "$dbdump_pluto_security" "$dbdump_pluto_telecom"
-#	rm -rf "${dbdump_dir}"
-
+#		rm -f "$dbdump_main_sqlcvs" "$dbdump_myth_sqlcvs" "$dbdump_pluto_media" "$dbdump_pluto_security" "$dbdump_pluto_telecom"
+#		rm -rf "${dbdump_dir}"
+	fi
 }
 
 DisplayMessage "*** STEP: Importing Databases"
